full linked push
here we have the entire functional structure of the website.

in the next push ill be pushing all base content to the website.

after that the push will be the full functionality push at which point the website will be done.

a final precautionary push will be done for minor grammar mistakes.